requirements.txt
google.generativeai
discord.py
fuzzywuzzy
python-dotenv
tenacity
python-Levenshtein
psutil